GH-1494 move Lucene default impl to org.eclipse.rdf4j.sail.lucene.impl
