+++ compliance/lucene/src/test/java/org/eclipse/rdf4j/sail/lucene/LuceneSailTest.java
@@ -15,6 +15,7 @@ import org.eclipse.rdf4j.common.iteration.Iterations;
 import org.eclipse.rdf4j.query.BindingSet;
 import org.eclipse.rdf4j.repository.RepositoryConnection;
 import org.eclipse.rdf4j.repository.util.Repositories;
+import org.eclipse.rdf4j.sail.lucene.impl.LuceneIndex;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -27,7 +28,7 @@ public class LuceneSailTest extends AbstractLuceneSailTest {
 	/*
 	 * (non-Javadoc)
 	 *
-	 * @see org.eclipse.rdf4j.sail.lucene.AbstractLuceneSailTest#configure(org.eclipse.rdf4j.sail.lucene.LuceneSail)
+	 * @see org.eclipse.rdf4j.sail.lucene.impl.AbstractLuceneSailTest#configure(org.eclipse.rdf4j.sail.lucene.impl.LuceneSail)
 	 */
 	@Override
 	protected void configure(LuceneSail sail) throws IOException {
