+++ core/sail/lucene/src/main/resources/META-INF/services/org.eclipse.rdf4j.sail.config.SailFactory
@@ -1,1 +1,1 @@
-org.eclipse.rdf4j.sail.lucene.config.LuceneSailFactory
+org.eclipse.rdf4j.sail.lucene.impl.config.LuceneSailFactory
