+++ site/content/release-notes/4.0.0.md
@@ -63,6 +63,10 @@ Both options may require shutting down the application(s) using the indexed data
 
 The RDF4J SDK contains both the `lucene-core` and `lucene-backwards-codec` jars.
 
+### Lucene default SearchIndex implementation renamed 
+
+The default Lucene `SearchIndex` implementation has been renamed. Prevously, it was `org.eclipse.rdf4j.sail.lucene.LuceneIndex`. The new name is `org.eclipse.rdf4j.sail.lucene.impl.LuceneIndex`. 
+
 ### Solr client libraries upgraded
 
 The Solr client libraries have been upgraded from 7.7 to 8.4.
